QC/test_LRU.py
import time
import pytest
from LRU import GeoDistributedLRUCache

# Simulate cache instance for testing
@pytest.fixture
def cache():
    return GeoDistributedLRUCache("example_cache:", 3600, ["redis1.example.com", "redis2.example.com"])

# Simplicity: Integration needs to be dead simple.
def test_integration_simplicity(cache):
    cache.set("user:123", "JohnDoe")
    result = cache.get("user:123")
    assert result == "JohnDoe"

# Resilient to network failures or crashes.
def test_network_failure_resilience(cache):
    # Simulate network failure during cache access
    cache.redis_clients[0].ping = lambda: False  # Mocking a network failure
    result = cache.get("user:123")
    assert result is None

# Near real-time replication of data across Geolocation.
def test_real_time_replication(cache):
    cache.set("user:123", "JohnDoe")
    time.sleep(1)  # Simulate near real-time replication
    result = cache.redis_clients[1].get("example_cache:user:123")
    assert result == "JohnDoe"

# Data consistency across regions.
def test_data_consistency_across_regions():
    cache_region1 = GeoDistributedLRUCache("example_cache:", 3600, ["redis1.example.com"])
    cache_region2 = GeoDistributedLRUCache("example_cache:", 3600, ["redis2.example.com"])
    cache_region1.set("user:123", "JohnDoe")
    result = cache_region2.get("user:123")
    assert result == "JohnDoe"

# Locality of reference.
def test_locality_of_reference():
    cache_region1 = GeoDistributedLRUCache("example_cache:", 3600, ["redis1.example.com"])
    cache_region2 = GeoDistributedLRUCache("example_cache:", 3600, ["redis2.example.com"])
    cache_region1.set("user:123", "JohnDoe")
    result = cache_region2.get("user:123")
    assert result == "JohnDoe"

# Flexible Schema.
def test_flexible_schema(cache):
    cache.set("int_key", 42)
    cache.set("str_key", "Hello, World!")
    cache.set("list_key", [1, 2, 3])
    int_result = cache.get("int_key")
    str_result = cache.get("str_key")
    list_result = cache.get("list_key")
    assert int_result == 42
    assert str_result == "Hello, World!"
    assert list_result == [1, 2, 3]

# Cache can expire.
def test_cache_expiration():
    cache = GeoDistributedLRUCache("example_cache:", 2, ["redis1.example.com"])
    cache.set("user:123", "JohnDoe")
    time.sleep(3)  # Wait for TTL to expire
    result = cache.get("user:123")
    assert result is None
